day7/bluetooth-on-esp/bt-control/pyproject.toml
[project]
name = "bt-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pyserial>=3.5",
]
